anything that uses it can handle a double, #437

+++ src/Player.h
@@ -32,7 +32,7 @@ public:
 	enum FlightControlState { CONTROL_MANUAL, CONTROL_FIXSPEED, CONTROL_AUTOPILOT };
 	FlightControlState GetFlightControlState() const { return m_flightControlState; }
 	void SetFlightControlState(FlightControlState s);
-	float GetSetSpeed() const { return m_setSpeed; }
+	double GetSetSpeed() const { return m_setSpeed; }
 	virtual bool OnDamage(Object *attacker, float kgDamage);
 	virtual void OnHaveKilled(Body *guyWeKilled);
 	int GetKillCount() const { return m_knownKillCount; }
